Make DocumentId handling only relevant to Collection FilterClauses
